+++ libraries/pizzaMenu.py
@@ -1,6 +1,6 @@
 import tkinter as tk
-from libraries.utils import capfirst
-from libraries.windowSpecs import WindowSpecs
+from libraries.utils.utils import capfirst
+from libraries.utils.windowSpecs import WindowSpecs
 from libraries.cells.gridTitleCell import TitleCell
 from libraries.cells.gridPizzaCell import PizzaCell
 from libraries.cells.gridAggiuntaCell import AggiuntaCell
