+++ libraries/utils/windowSpecs.py
