+++ pizzaMenu1.py
@@ -1,5 +1,5 @@
 from libraries.pizzaHandler import Pizzas
-from libraries.utils import waitForConnection
+from libraries.utils.utils import waitForConnection
 from libraries.verticalGrid import VerticalGrid
 from libraries.pizzaMenu import PizzaMenu
